RuoYi-Vue/report-gen/report-orchestrator/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
rank_bm25
nltk
sentence-transformers
faiss-cpu
psycopg[binary]
sqlalchemy[asyncio]
asyncmy
